requirements.txt
black==22.8.0
factory_boy
datetime==4.6
django==3.2
django-filter==22.1
djangorestframework==3.13.1
factory-boy==3.2.1
numpy==1.21.6
pandas==1.3.5
